pyproject.toml
[project]
name = "sqlalchemyapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "contextvars-executor>=0.0.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sadisplay>=0.4.9",
    "sqlalchemy>=2.0.45",
]
